Manual report creation

+++ app/domain/models/report_History.py
@@ -6,7 +6,7 @@ class ReportHistory(db.Model):
     __tablename__ = "report_history"
 
     id = db.Column(db.Integer, primary_key=True)
-    report_id = db.Column(db.Integer, db.ForeignKey("reports.id"), nullable=False, index=True)
+    report_id = db.Column(db.BigInteger, db.ForeignKey("reports.id"), nullable=False, index=True)
     status = db.Column(db.Enum("nuevo", "en_proceso", "finalizado", "archivado"), nullable=False)
     comment = db.Column(db.Text, nullable=True)
     changed_by_user_login = db.Column(db.String(100), nullable=True)
